Logging: update global log first.
If global logging is enabled, write to that log before the standard
log.

This ordering guards against the scenario where if the runtime misbehaves
somehow and the container is then deleted, the container state
directory will be deleted. But if the runtime is still in fact running,
it can no longer log to the standard log (ENOENT), but that write
failure means the global log will not be written to either.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>